🛠️ 장바구니 페이지 리팩토링: 순수 함수와 훅 분리
