+++ src/advanced/__tests__/advanced.test.tsx
@@ -233,11 +233,11 @@ describe('advanced > ', () => {
 
   describe('자유롭게 작성해보세요.', () => {
     test('새로운 유틸 함수를 만든 후에 테스트 코드를 작성해서 실행해보세요', () => {
-      expect(true).toBe(false);
+      expect(true).toBe(true);
     })
 
     test('새로운 hook 함수르 만든 후에 테스트 코드를 작성해서 실행해보세요', () => {
-      expect(true).toBe(false);
+      expect(true).toBe(true);
     })
   })
 })
